Extract Takens embedding optimization logic into a separate function
